add first three tasks

+++ MySiteSolution/Catalog.BLL/Models/Product.cs
@@ -9,7 +9,7 @@
         public Category Category { get; set; }
 
         public string QuantityPerUnit { get; set; }
-        public short UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public short UnitsInStock { get; set; }
         public short UnitsOnOrder { get; set; }
         public short ReorderLevel { get; set; }
